backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-CORS==4.0.0
PyMongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
openai==1.12.0
Werkzeug==3.0.1
marshmallow==3.21.0
Pillow==10.2.0
python-slugify==8.0.4
langdetect==1.0.9
gunicorn==21.2.0
pytest==7.4.4
